Type: docs Scope: Not applicable Commit Message: docs: Add MIT License to the project
This commit adds the MIT License to the project. The license grants permission to any person obtaining a copy of this software and associated documentation files (the “Software”), to deal in the Software without restriction, including without limitation the rights to use, copy, modify, merge, publish, distribute, sublicense, and/or sell copies of the Software. The license also includes a disclaimer of warranty.